pyproject.toml
[project]
name = "chimerapy-engine"
version = "0.1.0"
description = "ChimeraPy-Engine: The Cluster Networking Engine for the ChimeraPy Framework"
authors = [
    {name = "Eduardo Davalos", email="eduardo.davalos.anaya@vanderbilt.edu"},
    {name = "Umesh Timalsina", email="umesh.timalsina@vanderbilt.edu"}
]
license = {file = "LICENSE.txt"}
readme = "README.md"
requires-python = ">3.6"

keywords = ["multimodal", "data", "distributed", "streaming", "real-time"]

classifiers = [
    "Programming Language :: Python :: 3"
]

dependencies = [
    'networkx',
    'dill',
    'matplotlib',
    'multiprocess',
    'opencv-python',
    'pyaudio',
    'aioshutil',
    'asyncio-atexit',
    'pandas',
    'tqdm',
    'pyzmq',
    'simplejpeg',
    'aiohttp',
    'blosc',
    'PyYAML',
    'dataclasses-json',
    'aiofiles',
    'zeroconf',
    'aioreactive',
    'psutil'
]

[project.optional-dependencies]
test = [
    'pytest',
    'pytest-asyncio',
    'pytest-repeat',
    'pytest-lazy-fixture',
    'pytest-order',
    'pytest-rerunfailures',
    'pytest-timeout',
    'auto-changelog',
    'coveralls',
    'pre-commit',
    'docker',
    'numpy',
    'imutils',
    'pillow',
    'bump2version'
]
types = [
    'types-PyYAML',
    'types-aiofiles'
]
docs = [
    'sphinx',
    'myst-parser'
]
benchmark = [
    'lz4',
    'nujson',
    'orjson',
    'cloudpickle',
    'msgpack',
    'msgpack_numpy'
]
examples = [
    'dxcam; sys_platform == "win32"',
    'imutils',
    'numpy',
    'pillow'
]

[project.urls]
homepath = "https://github.com/ChimeraPy"
documentation = "https://chimerapy-engine.readthedocs.io/en/latest/index.html"
repository = "https://github.com/ChimeraPy/Engine"

# Entrypoint
[project.scripts]
cp-worker = "chimerapy.engine.entry.worker:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

# https://setuptools.pypa.io/en/stable/userguide/datafiles.html
[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
chimerapy = ['engine/*.yaml']

[tool.mypy]
files = "chimerapy"
ignore_missing_imports = true

# Reference:
# https://stackoverflow.com/questions/4673373/logging-within-pytest-tests

[tool.pytest.ini_options]
asyncio_mode = 'auto'

# Logging + CLI
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s.%(msecs)03d [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

# Logging + File
log_file = "chimerapy-engine-test.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s.%(msecs)03d [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_file_date_format = "%Y-%m-%d %H:%M:%S"

# Timeout
faulthandler_timeout=300

# Ignore warnings
filterwarnings = "ignore::DeprecationWarning"

markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit: marks tests as unit tests (deselect with '-m \"not unit\"')"
]
